backend/requirements.txt
absl-py==0.14.1
anaconda-client==1.7.2
aniso8601==8.0.0
appdirs==1.4.4
argh==0.26.2
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
beautifulsoup4==4.6.0
bkcharts==0.2
black==19.10b0
blinker==1.4
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.5.30
chardet==3.0.4
clang==5.0
clyent==1.2.2
comtypes==1.1.9
contextlib2==0.6.0.post1
cycler==0.10.0
cytoolz==0.11.0
django-allauth==0.45.0
django-cors-headers==3.9.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.0.1
executing==0.8.2
fastcache==1.1.0
Flask-RESTful==0.3.8
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gmpy2==2.0.8
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlemaps==4.5.3
grpcio==1.41.0
h5py==3.1.0
HeapDict==1.0.1
icecream==2.1.1
jdcal==1.4.1
Jinja2==2.11.2
JPype1==1.0.2
json5==0.9.5
jsonify==0.5
jupyter==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
konlpy==0.5.2
llvmlite==0.36.0
locket==0.2.1
LunarCalendar==0.0.9
MarkupSafe==1.1.1
mccabe==0.6.1
menuinst==1.4.16
mglearn==0.1.9
mistune==0.8.4
mkl-fft==1.3.0
mkl-service==2.3.0
mpmath==1.2.1
multipledispatch==0.6.0
multitasking==0.0.9
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.58
opt-einsum==3.3.0
pandas-datareader==0.9.0
parso==0.7.0
pathspec==0.7.0
patsy==0.5.1
pep8==1.7.1
pkginfo==1.7.0
plotly==5.0.0
ply==3.11
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycurl==7.43.0.6
pyodbc===4.0.0-unsupported
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyreadline==2.1
pystan==2.19.1.1
pytest==6.2.3
python3-openid==3.2.0
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
selenium==3.141.0
simplegeneric==0.8.1
sip==4.19.13
SQLAlchemy==1.3.20
statsmodels==0.12.2
tables==3.6.1
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.4
torch==1.9.1
tweepy==3.10.0
unicodecsv==0.14.1
urllib3==1.25.10
webencodings==0.5.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
xlwings==0.23.0
xlwt==1.3.0
yfinance==0.1.64
zict==2.0.0
zope.event==4.5.0
